requirements.txt
python-socketio
python-socketio[client]
eventlet
keyboard
opencv-python